+++ evm/db/chain.py
@@ -461,49 +461,3 @@ class ChainDB(HeaderDB, BaseChainDB):
         """
         for key, value in trie_data_dict.items():
             self.db[key] = value
-
-
-class AsyncChainDB(ChainDB):
-    async def coro_get(self, key: bytes) -> bytes:
-        raise NotImplementedError()
-
-    async def coro_get_score(self, block_hash: Hash32) -> int:
-        raise NotImplementedError()
-
-    async def coro_get_block_header_by_hash(self, block_hash: Hash32) -> BlockHeader:
-        raise NotImplementedError()
-
-    async def coro_get_canonical_head(self) -> BlockHeader:
-        raise NotImplementedError()
-
-    async def coro_get_canonical_block_header_by_number(
-            self, block_number: BlockNumber) -> BlockHeader:
-        raise NotImplementedError()
-
-    async def coro_header_exists(self, block_hash: Hash32) -> bool:
-        raise NotImplementedError()
-
-    async def coro_get_canonical_block_hash(self, block_number: BlockNumber) -> Hash32:
-        raise NotImplementedError()
-
-    async def coro_persist_header(self, header: BlockHeader) -> Tuple[BlockHeader, ...]:
-        raise NotImplementedError()
-
-    async def coro_persist_uncles(self, uncles: Tuple[BlockHeader]) -> Hash32:
-        raise NotImplementedError()
-
-    async def coro_persist_trie_data_dict(self, trie_data_dict: Dict[bytes, bytes]) -> None:
-        raise NotImplementedError()
-
-    async def coro_get_block_transactions(
-            self,
-            header: BlockHeader,
-            transaction_class: Type['BaseTransaction']) -> Iterable['BaseTransaction']:
-        raise NotImplementedError()
-
-    async def coro_get_block_uncles(self, uncles_hash: Hash32) -> List[BlockHeader]:
-        raise NotImplementedError()
-
-    async def coro_get_receipts(
-            self, header: BlockHeader, receipt_class: Type[Receipt]) -> List[Receipt]:
-        raise NotImplementedError()
