+++ tests/p2p/integration_test_helpers.py
@@ -4,10 +4,10 @@ from evm import MainnetChain, RopstenChain
 from evm.chains.base import (
     MiningChain,
 )
-from evm.db.chain import AsyncChainDB
 
 from p2p.exceptions import OperationCancelled
 
+from trinity.db.chain import AsyncChainDB
 from trinity.db.header import AsyncHeaderDB
 
 
@@ -31,18 +31,20 @@ def async_passthrough(base_name):
     return passthrough_method
 
 
-class FakeAsyncChainDB(AsyncChainDB):
-    coro_get = async_passthrough('get')
-    coro_get_score = async_passthrough('get_score')
-    coro_get_block_header_by_hash = async_passthrough('get_block_header_by_hash')
-    coro_get_canonical_head = async_passthrough('get_canonical_head')
-    coro_header_exists = async_passthrough('header_exists')
+class FakeAsyncHeaderDB(AsyncHeaderDB):
     coro_get_canonical_block_hash = async_passthrough('get_canonical_block_hash')
+    coro_get_canonical_block_header_by_number = async_passthrough('get_canonical_block_header_by_number')  # noqa: E501
+    coro_get_canonical_head = async_passthrough('get_canonical_head')
+    coro_get_block_header_by_hash = async_passthrough('get_block_header_by_hash')
+    coro_get_score = async_passthrough('get_score')
+    coro_header_exists = async_passthrough('header_exists')
     coro_persist_header = async_passthrough('persist_header')
+
+
+class FakeAsyncChainDB(FakeAsyncHeaderDB, AsyncChainDB):
     coro_persist_uncles = async_passthrough('persist_uncles')
     coro_persist_trie_data_dict = async_passthrough('persist_trie_data_dict')
-    coro_get_canonical_block_header_by_number = async_passthrough(
-        'get_canonical_block_header_by_number')
+    coro_get = async_passthrough('get')
     coro_get_block_transactions = async_passthrough('get_block_transactions')
     coro_get_block_uncles = async_passthrough('get_block_uncles')
     coro_get_receipts = async_passthrough('get_receipts')
@@ -70,13 +72,3 @@ class FakeAsyncMainnetChain(MainnetChain):
 class FakeAsyncChain(MiningChain):
     coro_import_block = coro_import_block
     coro_validate_chain = async_passthrough('validate_chain')
-
-
-class FakeAsyncHeaderDB(AsyncHeaderDB):
-    coro_get_canonical_block_hash = async_passthrough('get_canonical_block_hash')
-    coro_get_canonical_block_header_by_number = async_passthrough('get_canonical_block_header_by_number')  # noqa: E501
-    coro_get_canonical_head = async_passthrough('get_canonical_head')
-    coro_get_block_header_by_hash = async_passthrough('get_block_header_by_hash')
-    coro_get_score = async_passthrough('get_score')
-    coro_header_exists = async_passthrough('header_exists')
-    coro_persist_header = async_passthrough('persist_header')
